+++ src/components/MatchCriterion/edit/ExistingRecordSections/ExistingSectionFolio.js
@@ -1,6 +1,6 @@
 import React, { useCallback } from 'react';
 import PropTypes from 'prop-types';
-import { Field } from 'redux-form';
+import { Field } from 'react-final-form';
 import classnames from 'classnames';
 import { last } from 'lodash';
 
@@ -28,7 +28,8 @@ export const ExistingSectionFolio = ({
   existingRecordFieldLabel,
   existingRecordFields,
   existingRecordType,
-  dispatchFormChange,
+  changeFormState,
+  formValues,
 }) => {
   const fieldToChangeName = `profile.matchDetails[${repeatableIndex}].existingMatchExpression.fields[0].value`;
 
@@ -36,7 +37,7 @@ export const ExistingSectionFolio = ({
     return dataOptions.filter(o => new RegExp(`${value}`, 'i').test(o.label));
   };
 
-  const handleExistingRecordSelect = newValue => {
+  const handleExistingRecordSelect = (newValue, fieldProps) => {
     const fieldToChange = `profile.matchDetails[${repeatableIndex}].existingMatchExpression.fields`;
     const fieldId = existingRecordFields.find(item => item.value === newValue)?.id;
     const fieldFromConfig = fieldsConfig.find(item => item.id === fieldId && item.recordType === existingRecordType);
@@ -54,25 +55,28 @@ export const ExistingSectionFolio = ({
       });
     }
 
-    dispatchFormChange(fieldToChange, fieldToChangeValue);
+    fieldProps.input.onChange(newValue);
+    changeFormState(fieldToChange, fieldToChangeValue);
   };
 
   const formatExistingRecordValue = useCallback(
     value => {
       if (!value) return '';
 
-      const fieldValue = value[0].value;
+      const valueFromState = formValues.profile.matchDetails[repeatableIndex].existingMatchExpression.fields;
+
+      const fieldValue = valueFromState[0].value;
       const fieldFromConfig = fieldsConfig.find(item => item.value === fieldValue && item.recordType === existingRecordType);
 
       if (fieldFromConfig?.fromResources) {
-        return last(value).value;
+        return last(valueFromState).value;
       }
 
       const existingRecordField = existingRecordFields.find(item => item.id === fieldFromConfig?.id)?.value;
 
       return existingRecordField || '';
     },
-    [existingRecordFields, existingRecordType],
+    [existingRecordFields, existingRecordType, formValues.profile.matchDetails, repeatableIndex],
   );
 
   return (
@@ -84,14 +88,18 @@ export const ExistingSectionFolio = ({
         <Col xs={12}>
           <Field
             id="criterion-value-type"
-            aria-label={existingRecordFieldLabel}
-            component={Selection}
             name={fieldToChangeName}
-            dataOptions={existingRecordFields}
+            validate={validateRequiredField}
             format={formatExistingRecordValue}
-            onChange={handleExistingRecordSelect}
-            onFilter={handleFieldSearch}
-            validate={[validateRequiredField]}
+            render={criterionFieldProps => (
+              <Selection
+                {...criterionFieldProps}
+                aria-label={existingRecordFieldLabel}
+                dataOptions={existingRecordFields}
+                onFilter={handleFieldSearch}
+                onChange={value => handleExistingRecordSelect(value, criterionFieldProps)}
+              />
+            )}
           />
         </Col>
       </Row>
@@ -102,12 +110,13 @@ export const ExistingSectionFolio = ({
 ExistingSectionFolio.propTypes = {
   repeatableIndex: PropTypes.number.isRequired,
   existingRecordType: PropTypes.oneOf([...Object.keys(FOLIO_RECORD_TYPES), '']).isRequired,
-  dispatchFormChange: PropTypes.func.isRequired,
+  changeFormState: PropTypes.func.isRequired,
   existingRecordFields: PropTypes.arrayOf(PropTypes.shape({
     value: PropTypes.string.isRequired,
     label: PropTypes.string.isRequired,
   })),
   existingRecordFieldLabel: PropTypes.string,
+  formValues: PropTypes.object,
 };
 
 ExistingSectionFolio.defaultProps = {